Set up project to be used in container via VS Code
Using a language is sometimes as much about the IDE experience as the syntax/grammar/etc. So I'm going to develop Rithm in a way that is tightly coupled to VS Code, since it is easy and easy to modify. Also, it means I can set up the dev environment in a docker container, and avoid a lot of the problems of bootstrapping a dev environment (hopefully)